requirements.txt
python-telegram-bot==22.7
aiohttp==3.13.5
python-dotenv==1.2.2